requirements.txt
python-dotenv
streamlit
langchain-experimental
openai
tabulate
pandas